Refugees by UNHCR
By using Excel with all its analysis tools such as Power Query, Pivot Table, and Tableau with many other visualization software. Throughout the study, we started by finding a correlation between Four of our variables which are Sum of refugees, the sum of total refugees and people in refugee-like situations, the sum of refugees assisted by UNHCR, and the sum of total refugees and people in refugee-like situations assisted by UNHCR.
When these links are established, we will eliminate the low correlation variables. To be able to write a linear regression. Then, the main task is to tackle the high rate of refugees and their country of origin. There are refugees in certain countries or continents that are less assisted by UNHCR. The study will show the high flow of refugees from 1975 to 2021 across the world.